Move codeowner code around
- Move codeowner related code out of the models package, as they quite
frankly don't have anything to do with database models. It has been
moved to `modules/codeowner` for the codeowner file parsing and
`services/issue` for the assinging of codeowners to pull requests.

Refs: https://codeberg.org/forgejo/forgejo/pulls/1970

(cherry picked from commit ea6942585d43dba8b038758782631e58ea8a34b4)
