+++ models/issues/pull_test.go
@@ -306,28 +306,6 @@ func TestDeleteOrphanedObjects(t *testing.T) {
 	assert.EqualValues(t, countBefore, countAfter)
 }
 
-func TestParseCodeOwnersLine(t *testing.T) {
-	type CodeOwnerTest struct {
-		Line   string
-		Tokens []string
-	}
-
-	given := []CodeOwnerTest{
-		{Line: "", Tokens: nil},
-		{Line: "# comment", Tokens: []string{}},
-		{Line: "!.* @user1 @org1/team1", Tokens: []string{"!.*", "@user1", "@org1/team1"}},
-		{Line: `.*\\.js @user2 #comment`, Tokens: []string{`.*\.js`, "@user2"}},
-		{Line: `docs/(aws|google|azure)/[^/]*\\.(md|txt) @org3 @org2/team2`, Tokens: []string{`docs/(aws|google|azure)/[^/]*\.(md|txt)`, "@org3", "@org2/team2"}},
-		{Line: `\#path @org3`, Tokens: []string{`#path`, "@org3"}},
-		{Line: `path\ with\ spaces/ @org3`, Tokens: []string{`path with spaces/`, "@org3"}},
-	}
-
-	for _, g := range given {
-		tokens := issues_model.TokenizeCodeOwnersLine(g.Line)
-		assert.Equal(t, g.Tokens, tokens, "Codeowners tokenizer failed")
-	}
-}
-
 func TestGetApprovers(t *testing.T) {
 	assert.NoError(t, unittest.PrepareTestDatabase())
 	pr := unittest.AssertExistsAndLoadBean(t, &issues_model.PullRequest{ID: 5})
